refactor: reorganize format functions
